+++ src/shell/components/global-sidebar/components/InstanceMenu/DropdownMenu.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const DropdownMenu = () => {
+  return <div>DropdownMenu</div>;
+};
